+++ doc/whats-new.rst
@@ -35,7 +35,7 @@ Breaking changes
 
 Deprecations
 ~~~~~~~~~~~~
-
+xarray.cftime_range deprecated in favor of xarray.data_range(use_cftime=true)
 
 Bug fixes
 ~~~~~~~~~
@@ -48,6 +48,9 @@ Bug fixes
 - Use mean of min/max years as offset in calculation of datetime64 mean
   (:issue:`10019`, :pull:`10035`).
   By `Kai Mühlbauer <https://github.com/kmuehlbauer>`_.
+- Deprecate xr.cftime_range() in favor of xr.date_range(use_cftime=True)
+  (:issue:`9886`, :pull:`10024`).
+  By `Josh Kihm <https://github.com/maddogghoek>`_.
 - Fix DataArray().drop_attrs(deep=False) and add support for attrs to
   DataArray()._replace(). (:issue:`10027`, :pull:`10030`). By `Jan
   Haacker <https://github.com/j-haacker>`_.